refactor, add diff cmd
